backend/requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
qrcode==8.2
